requirements.txt
aiodataloader==0.4.3
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
ariadne==0.26.2
asgiref==3.11.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==5.0.0
billiard==4.2.4
blinker==1.9.0
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.12.0
django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
faker==37.12.0
fastapi==0.121.3
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.0
fastar==0.6.0
flask==3.1.2
flask-sqlalchemy==3.1.1
graphql-core==3.2.5
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
jinja2==3.1.6
kombu==5.6.1
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
prompt-toolkit==3.0.52
psycopg2-binary==2.9.9
pwdlib==0.3.0
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pygments==2.19.2
pyjwt==2.10.1
pytest==9.0.1
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pyyaml==6.0.3
redis==7.1.0
requests==2.32.5
rich==14.2.0
rich-toolkit==0.16.0
rignore==0.7.6
sentry-sdk==2.45.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.44
sqlmodel==0.0.27
sqlparse==0.5.3
starlette==0.50.0
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
werkzeug==3.1.3
